version final

+++ controladores/usuarios.controlador.php
@@ -505,7 +505,7 @@
 						/* Creamos el directorio */
 						$directorio = "vistas/img/usuarios/".$_POST['EditarCorreo'];	
 						/* Primero preguntamos su existe una imagen en la base de datos */
-						if(!empty($_POST['fotoActual'])){
+						if(!empty($_POST['fotoActual']) && is_file($_POST['fotoActual'])){
 							unlink($_POST['fotoActual']);
 						}else{
 							mkdir($directorio, 0755);
